change all doubles to floats

+++ src/lemlib/chassis/chassis.cpp
@@ -75,7 +75,7 @@ void lemlib::Chassis::calibrate() {
  * @param theta new theta value
  * @param radians true if theta is in radians, false if not. False by default
  */
-void lemlib::Chassis::setPose(double x, double y, double theta, bool radians) {
+void lemlib::Chassis::setPose(float x, float y, float theta, bool radians) {
     lemlib::setPose(lemlib::Pose(x, y, theta), radians);
 }
